% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basketball_births}
\alias{basketball_births}
\title{The birth dates of NBA/ABA players}
\format{
A data frame with columns:
\describe{
 \item{player}{The variable is character. Player's name}
 \item{birthday}{The variable is Date. Player's birthdate}
 \item{yrs}{The variable is integer. Number of years played}
}
}
\source{
\url{https://www.basketball-reference.com/friv/birthdays.fcgi?month=1&day=1}
}
\usage{
basketball_births
}
\description{
Data obtained from https://www.basketball-reference.com
}
\examples{
\dontrun{
basketball_births
}
}
\keyword{datasets}
