% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_all}
\alias{counts_all}
\title{The count of births of all players and US by month.}
\format{
A data frame with columns:
\describe{
 \item{month_number}{The variable is numeric. The numeric month of the year}
 \item{month_name}{The variable is character. The name of the month}
 \item{day_of_month}{The variable is numeric. The day number of the month 1-31}
 \item{day_of_year}{The variable is integer. The day number of the year 1-366}
 \item{births}{The variable is numeric. The total count of births on that day over the period of the data}
 \item{group}{The variable is character. The professional sport or US births}
}
}
\source{
\url{https://github.com/byuidatascience/data4births}
}
\usage{
counts_all
}
\description{
Data combined from others sources in package.
}
\examples{
\dontrun{
counts_all
}
}
\keyword{datasets}
