% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_baseball}
\alias{counts_baseball}
\title{The count of births of MLB players for US born players from 1925 to 2015}
\format{
A data frame with columns:
\describe{
 \item{month_number}{The variable is integer. The numeric month of the year}
 \item{month_name}{The variable is character. The name of the month}
 \item{day_of_month}{The variable is integer. The day number of the month 1-31}
 \item{day_of_year}{The variable is integer. The day number of the year 1-366}
 \item{births}{The variable is integer. The total count of births on that day over the period of the data}
}
}
\source{
\url{https://github.com/cdalzell/Lahman}
}
\usage{
counts_baseball
}
\description{
Data obtained from Lahman http://www.seanlahman.com/baseball-archive/statistics/
}
\examples{
\dontrun{
counts_baseball
}
}
\keyword{datasets}
