% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_us}
\alias{counts_us}
\title{The count of births in the United States from 1994-2014}
\format{
A data frame with columns:
\describe{
 \item{month_number}{The variable is numeric. The numeric month of the year}
 \item{month_name}{The variable is character. The name of the month}
 \item{day_of_month}{The variable is numeric. The day number of the month 1-31}
 \item{day_of_year}{The variable is integer. The day number of the year 1-366}
 \item{births}{The variable is numeric. The total count of births on that day over the period of the data}
}
}
\source{
\url{https://github.com/fivethirtyeight/data/tree/master/births}
}
\usage{
counts_us
}
\description{
Data obtained from the CDC and Census parsed by FiveThirtyEight
}
\examples{
\dontrun{
counts_us
}
}
\keyword{datasets}
