% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football_births}
\alias{football_births}
\title{The birth dates of NFL players}
\format{
A data frame with columns:
\describe{
 \item{player}{The variable is character. Player's name}
 \item{birthday}{The variable is Date. Player's birthdate}
 \item{yrs}{The variable is integer. Number of years played}
 \item{from}{The variable is integer. Year started}
 \item{to}{The variable is integer. Year ended}
 \item{pos}{The variable is character. Positions played}
}
}
\source{
\url{https://www.pro-football-reference.com/years/1900/births.htm}
}
\usage{
football_births
}
\description{
Data obtained from https://www.pro-football-reference.com
}
\examples{
\dontrun{
football_births
}
}
\keyword{datasets}
