% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hockey_births}
\alias{hockey_births}
\title{The birth dates of NHL players}
\format{
A data frame with columns:
\describe{
 \item{player}{The variable is character. Player's name}
 \item{birthday}{The variable is Date. Player's birthdate}
 \item{yrs}{The variable is integer. Number of years played}
 \item{from}{The variable is integer. Year started}
 \item{to}{The variable is integer. Year ended}
 \item{team}{The variable is character. Teams played with}
}
}
\source{
\url{https://www.hockey-reference.com/friv/birthdays.cgi?month=1&day=1}
}
\usage{
hockey_births
}
\description{
Data obtained from https://www.hockey-reference.com
}
\examples{
\dontrun{
hockey_births
}
}
\keyword{datasets}
